CHI-137 (CHI-11 subtask 3): replace Spring Mat3x3d members with fp32 storage
Migrates Spring::dp_dx1, dp_dx2 from Eigen::Matrix<double,3,3> to
Eigen::Matrix<float,3,3> as the constraint-matrix-member migration
pilot. Adds a scalar-templated insertIntoTriplets overload in
UtilityFunctions.h so triplet emission accepts any element type and
casts at the boundary, removing the need for per-call-site
.cast<double>() shims.

The previous CHI-11 attempts at subtask 3 failed because Eigen's
templated operator* forbids implicit float×double matrix products at
overload-resolution time. The fix used here is to keep storage fp32
and add an explicit .cast<float>() on the assignment of an fp64
expression to the fp32 cached buffer — the cast is value-level, so
template deduction succeeds.

Spring is a low-risk pilot: dp_dx1/dp_dx2 are written in
projectBackwardPrecompute and consumed only via insertIntoTriplets,
so the precision change does not leak into other constraint math.

+++ src/code/simulation/Spring.cpp
@@ -123,8 +123,8 @@ void Spring::projectBackwardPrecompute(const VecXd &x_vec) {
 
 	// TODO: no stiffness, is this right?
 	Mat3x3d ddir_dposdiff = (I_three - dir * dir.transpose()) / l;
-	dp_dx1 = l0 * ddir_dposdiff * d_posdiff_dx1;
-	dp_dx2 = l0 * ddir_dposdiff * d_posdiff_dx2;
+	dp_dx1 = (l0 * ddir_dposdiff * d_posdiff_dx1).cast<float>();
+	dp_dx2 = (l0 * ddir_dposdiff * d_posdiff_dx2).cast<float>();
 }
 
 void Spring::projectBackward(const VecXd &x_vec, TripleVector &triplets) {
